ADD waiting CRUD
Former-commit-id: 7bc937b76aafa0706801b33df24c2ad3f218c38b

+++ backend/src/main/java/com/bigdata/tikitacar/car/controller/DealController.java
@@ -2,6 +2,7 @@ package com.bigdata.tikitacar.car.controller;
 
 import com.bigdata.tikitacar.car.dto.request.DealRegisterRequestDto;
 import com.bigdata.tikitacar.car.dto.request.DealUpdateRequestDto;
+import com.bigdata.tikitacar.car.dto.request.DealUpdateStatusRequestDto;
 import com.bigdata.tikitacar.car.dto.response.DealSearchResponseDto;
 import com.bigdata.tikitacar.car.service.DealService;
 import com.bigdata.tikitacar.user.dto.response.UserFindResponseDto;
@@ -40,13 +41,23 @@ public class DealController {
         ResponseEntity response = null;
         Map<String,Object> map = new HashMap<String, Object>();
 
-        //유저정보 dto에 등록
-        dealRegisterRequestDto.updateSellerId(userService.findUserByEmail(jwtService.getEmailFromToken(token)).getId());
-        dealService.registerDeal(dealRegisterRequestDto);
+        String loginEamil = jwtService.getEmailFromToken(token);
 
-        map.put("msg","거래 등록에 성공했습니다.");
-        map.put("status","success");
-        response = new ResponseEntity(map, HttpStatus.OK);
+        if(loginEamil!=null){
+            //유저정보 dto에 등록
+            dealRegisterRequestDto.updateSellerId(userService.findUserByEmail(loginEamil).getId());
+            dealService.registerDeal(dealRegisterRequestDto);
+
+            map.put("msg","거래 등록에 성공했습니다.");
+            map.put("status","success");
+            response = new ResponseEntity(map, HttpStatus.OK);
+        }else{
+            map.put("msg","거래 등록에 실패했습니다.");
+            map.put("status","fail");
+            response = new ResponseEntity(map, HttpStatus.BAD_REQUEST);
+        }
+
+
         return response;
     }
 
@@ -80,7 +91,7 @@ public class DealController {
         String loginEmail = jwtService.getEmailFromToken(token);
         DealSearchResponseDto dealSearchResponseDto = dealService.searchDeal(dealUpdateRequestDto.getId());
 
-        if(loginEmail.equals(dealSearchResponseDto.getEmail())){
+        if(loginEmail!= null && loginEmail.equals(dealSearchResponseDto.getEmail())){
             dealService.updateDeal(dealUpdateRequestDto);
             map.put("msg","거래 수정에 성공했습니다.");
             map.put("status","success");
@@ -97,16 +108,25 @@ public class DealController {
     @ApiOperation("거래 완료처리")
     @PutMapping("/status/{id}")
     public Object dealDone(@RequestHeader(value="Authorization") String token,
-                           @PathVariable Long dealId){
+                           @PathVariable("id") Long dealId,
+                           @RequestBody DealUpdateStatusRequestDto dealUpdateStatusRequestDto){
         ResponseEntity response = null;
         Map<String,Object> map = new HashMap<String, Object>();
 
         //현재 로그인한 아이디
-        Long buyerId = userService.findUserByEmail(jwtService.getEmailFromToken(token)).getId();
-        dealService.updateDealStatus(buyerId,dealId);
+        String loginEmail = jwtService.getEmailFromToken(token);
+        DealSearchResponseDto dealSearchResponseDto = dealService.searchDeal(dealId);
 
-
-
+        if(loginEmail!=null && loginEmail.equals(dealSearchResponseDto.getEmail())){
+            dealService.updateDealStatus(dealId,dealUpdateStatusRequestDto);
+            map.put("msg","거래 완료에 성공했습니다.");
+            map.put("status","success");
+            response = new ResponseEntity(map,HttpStatus.OK);
+        }else{
+            map.put("msg","거래 완료에 실패했습니다.");
+            map.put("status","fail");
+            response = new ResponseEntity(map,HttpStatus.BAD_REQUEST);
+        }
         return response;
     }
 
@@ -122,7 +142,7 @@ public class DealController {
         String loginEmail = jwtService.getEmailFromToken(token);
         DealSearchResponseDto dealSearchResponseDto = dealService.searchDeal(id);
 
-        if(loginEmail.equals(dealSearchResponseDto.getEmail())){
+        if(loginEmail!= null && loginEmail.equals(dealSearchResponseDto.getEmail())){
             dealService.removeDeal(id);
             map.put("msg","거래 삭제에 성공했습니다.");
             map.put("status","success");
